+++ backend/python_scripts/marketDataCalcs.py
@@ -1,0 +1,52 @@
+# Pre-defined Libraries
+import math
+import json
+
+# Defined Libraries
+import initializeVariables
+import sessionMethods
+
+
+# Initialize Variables
+columnsWanted,number,step,stock,urls,headers = initializeVariables.initVars()
+
+# Method to get nearest strikes
+def round_nearest(x,num=50): return int(math.ceil(float(x)/num)*num)
+def nearest_strike_bnf(x): return round_nearest(x,step["bnf"])
+def nearest_strike_nf(x): return round_nearest(x,step["nf"])
+
+# Finding highest Open Interest of People's in CE based on CE data         
+def highest_oi_CE(num,step,nearest,url):
+    strike = nearest - (step*num)
+    start_strike = nearest - (step*num)
+    response_text = sessionMethods.get_data_with_retry(url)
+    data = json.loads(response_text)
+    currExpiryDate = data["records"]["expiryDates"][0]
+    max_oi = 0
+    max_oi_strike = 0
+    for item in data['records']['data']:
+        if item["expiryDate"] == currExpiryDate:
+            if item["strikePrice"] == strike and item["strikePrice"] < start_strike+(step*num*2):
+                if item["CE"]["openInterest"] > max_oi:
+                    max_oi = item["CE"]["openInterest"]
+                    max_oi_strike = item["strikePrice"]
+                strike = strike + step
+    return max_oi_strike
+
+# Finding highest Open Interest of People's in PE based on PE data 
+def highest_oi_PE(num,step,nearest,url):
+    strike = nearest - (step*num)
+    start_strike = nearest - (step*num)
+    response_text = sessionMethods.get_data_with_retry(url)
+    data = json.loads(response_text)
+    currExpiryDate = data["records"]["expiryDates"][0]
+    max_oi = 0
+    max_oi_strike = 0
+    for item in data['records']['data']:
+        if item["expiryDate"] == currExpiryDate:
+            if item["strikePrice"] == strike and item["strikePrice"] < start_strike+(step*num*2):
+                if item["PE"]["openInterest"] > max_oi:
+                    max_oi = item["PE"]["openInterest"]
+                    max_oi_strike = item["strikePrice"]
+                strike = strike + step
+    return max_oi_strike
